+++ tests/test_schema.py
@@ -50,10 +50,7 @@ def test_mlm_no_undefined_prefixed_field_item_properties(
     with pytest.raises(pystac.errors.STACValidationError) as exc:
         mlm_item = pystac.Item.from_dict(mlm_data)
         pystac.validation.validate(mlm_item, validator=mlm_validator)
-    assert all(
-        info in str(exc.value.source)
-        for info in ["mlm:unknown", "^(?!mlm:)"]
-    )
+    assert all(info in str(exc.value.source) for info in ["mlm:unknown", "^(?!mlm:)"])
 
     # defined property only allowed at the Asset level
     mlm_data = copy.deepcopy(mlm_example)
@@ -313,13 +310,13 @@ def test_mlm_asset_artifact_type_checked(
     mlm_item = pystac.Item.from_dict(mlm_data)
     pystac.validation.validate(mlm_item, validator=mlm_validator)  # self-check valid beforehand
 
-    mlm_data["assets"]["model"]["mlm:artifact_type"] = 1234  # type: ignore
+    mlm_data["assets"]["model"]["mlm:artifact_type"] = 1234
     mlm_item = pystac.Item.from_dict(mlm_data)
     with pytest.raises(pystac.errors.STACValidationError) as exc:
         pystac.validation.validate(mlm_item, validator=mlm_validator)
     assert "1234 is not of type 'string'" in str(exc.value.source)
 
-    mlm_data["assets"]["model"]["mlm:artifact_type"] = ""  # type: ignore
+    mlm_data["assets"]["model"]["mlm:artifact_type"] = ""
     mlm_item = pystac.Item.from_dict(mlm_data)
     with pytest.raises(pystac.errors.STACValidationError) as exc:
         pystac.validation.validate(mlm_item, validator=mlm_validator)